changed Gpa data type to double

+++ src/main/java/cmpt276/assn2/models/UserRepository.java
@@ -5,6 +5,6 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Integer> {
-    List<User> findByGpa(Integer gpa);
+    List<User> findByGpa(Double gpa);
     List<User> findByNameAndHeight(String name, Integer height);
 }
